Move inputs and outputs into fully qualified paths
This eliminates* the need to check for duplicates and paves the way for fancier
features such as wildcards.

It would be reasonable to just create a new root for each input/output
argument, but many existing genomics tools expect that files and their indices
are located in the same directory.  This should help reduce surprises for users
that expect things to just work as they would with a local directory.

* Technically, you could still get into trouble with something like
  --inputs=/foo/bar,gs://foo/bar