Fix bug when using actions and output-interval
When using both an actions list (from a JSON file) and the output-interval
flag, the background logging action is overwritten by the first action.

This causes two problems: (1) the logging action is gone and (2) the first
action may be mutated in unexpected ways since it is merged with the output
action.